+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/DailyViewModel.java
@@ -15,7 +15,7 @@ import snu.swpp.moment.data.model.StoryModel;
 import snu.swpp.moment.data.repository.AuthenticationRepository;
 import snu.swpp.moment.data.repository.MomentRepository;
 import snu.swpp.moment.data.repository.StoryRepository;
-import snu.swpp.moment.ui.main_writeview.DaySlide.StoryUiState;
+import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
 import snu.swpp.moment.ui.main_writeview.uistate.MomentUiState;
 import snu.swpp.moment.utils.EmotionMap;
 import snu.swpp.moment.utils.TimeConverter;
