+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/DaySlide/DailyViewFragment.java
@@ -29,6 +29,7 @@ import snu.swpp.moment.ui.main_writeview.DailyViewModelFactory;
 import snu.swpp.moment.ui.main_writeview.ListViewAdapter;
 import snu.swpp.moment.ui.main_writeview.ListViewItem;
 import snu.swpp.moment.ui.main_writeview.uistate.MomentUiState;
+import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
 import snu.swpp.moment.utils.TimeConverter;
 
 public class DailyViewFragment extends Fragment {
