+++ frontend/Moment/app/src/main/java/snu/swpp/moment/ui/main_writeview/DaySlide/ListFooterContainer.java
@@ -6,9 +6,9 @@ import androidx.annotation.NonNull;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.Observer;
-import java.util.Date;
 import snu.swpp.moment.R;
 import snu.swpp.moment.ui.main_writeview.EmotionGridContainer;
+import snu.swpp.moment.ui.main_writeview.uistate.StoryUiState;
 import snu.swpp.moment.utils.AnimationProvider;
 
 public class ListFooterContainer {
